In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int f1, f2, res, i;
+	unsigned long int f1, f2, res, i;
 
 	for (i = 0 ; i < 98 ; i++)
 	{
